+++ assets/js/main-0.1.3.js
@@ -426,34 +426,38 @@ document.addEventListener('DOMContentLoaded', function() {
   const maxDaysApart = 31;
   const today = new Date();
   const maxPastDate = new Date(today);
-  maxPastDate.setDate(today.getDate() - 45);
+  maxPastDate.setDate(today.getDate() - 31);
 
-  flatpickr("#startDate", {
-    enableTime: true,
-    dateFormat: "Y-m-d H:i",
+  flatpickr("#assessmentStartAndEndDate", {
+    mode: "range",
     minDate: maxPastDate,
     maxDate: today,
+    enableTime: true,
+    dateFormat: "Y-m-d H:i",
     onChange: function(selectedDates, dateStr, instance) {
-      const endDatePicker = document.getElementById('endDate')._flatpickr;
-      const maxEndDate = new Date(selectedDates[0]);
-      maxEndDate.setDate(maxEndDate.getDate() + maxDaysApart);
-      endDatePicker.set('minDate', dateStr);
-      endDatePicker.set('maxDate', maxEndDate > today ? today : maxEndDate);
+      if (selectedDates.length === 1) {
+        const startDate = selectedDates[0];
+        const maxEndDate = new Date(startDate.getTime() + 31 * 24 * 60 * 60 * 1000); // 31 days later
+        const today = new Date();
+        instance.set('maxDate', maxEndDate > today ? today : maxEndDate);
+      }
+      if (selectedDates.length === 2) {
+        const startDate = selectedDates[0];
+        const endDate = selectedDates[1];
+        const diffInDays = (endDate - startDate) / (1000 * 60 * 60 * 24);
+        if (diffInDays > 31) {
+          toast("Error","","The start and end date cannot exceed 31 days.","warning");
+          instance.clear();
+        }
+      }
     }
   });
 
-  flatpickr("#endDate", {
-    enableTime: true,
-    dateFormat: "Y-m-d H:i",
-    minDate: maxPastDate,
+  flatpickr("#reportingStartAndEndDate", {
+    mode: "range",
     maxDate: today,
-    onChange: function(selectedDates, dateStr, instance) {
-      const startDatePicker = document.getElementById('startDate')._flatpickr;
-      const minStartDate = new Date(selectedDates[0]);
-      minStartDate.setDate(minStartDate.getDate() - maxDaysApart);
-      startDatePicker.set('maxDate', dateStr);
-      startDatePicker.set('minDate', minStartDate < maxPastDate ? maxPastDate : minStartDate);
-    }
+    enableTime: true,
+    dateFormat: "Y-m-d H:i"
   });
 });